closed_contour/requirements.txt
absl-py==0.1.10
bleach==2.1.2
boltons==17.2.0
certifi==2020.4.5.1
chardet==3.0.4
click==6.7
cycler==0.10.0
decorator==4.4.2
eagerpy==0.27.0
entrypoints==0.2.3
foolbox==1.8.0
gitdb==4.0.5
GitPython==3.1.3
html5lib==1.0.1
idna==2.9
imageio==2.8.0
ipykernel==4.8.2
ipython==6.2.1
ipython-genutils==0.2.0
ipywidgets==7.1.2
jedi==0.11.1
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.2
jupyter-console==5.2.0
jupyter-contrib-core==0.3.3
jupyter-contrib-nbextensions==0.4.0
jupyter-core==4.4.0
jupyter-highlight-selected-word==0.1.0
jupyter-latex-envs==1.4.3
jupyter-nbextensions-configurator==0.4.0
jupyterlab==0.31.8
jupyterlab-launcher==0.10.5
kiwisolver==1.2.0
Lasagne==0.2.dev1
lxml==4.1.1
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==3.2.1
mistune==0.8.3
mypy==0.560
nbconvert==5.3.1
nbformat==4.4.0
networkx==2.4
notebook==5.4.0
numpy==1.18.4
pandas==0.22.0
pandocfilters==1.4.2
parso==0.1.1
pexpect==4.4.0
pickleshare==0.7.4
Pillow==7.1.2
prompt-toolkit==1.0.15
protobuf==3.12.2
psutil==5.4.3
ptyprocess==0.5.2
pycurl==7.43.0
Pygments==2.2.0
pygobject==3.20.0
pyparsing==2.4.7
python-apt==1.1.0b1
python-dateutil==2.8.1
pytz==2018.3
PyWavelets==1.1.1
PyYAML==3.12
pyzmq==17.0.0
qtconsole==4.3.1
requests==2.23.0
scikit-image==0.17.2
scikit-learn==0.19.1
scipy==1.4.1
seaborn==0.8.1
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.15.0
sklearn==0.0
smmap==3.0.4
tensorboardX==2.0
tensorflow-gpu==1.5.0
tensorflow-tensorboard==1.5.1
terminado==0.8.1
testpath==0.3.1
Theano==1.0.1
tifffile==2020.5.30
torch==0.4.0
torchvision==0.2.1
tornado==4.5.3
tqdm==4.19.5
traitlets==4.3.2
typed-ast==1.1.0
typing-extensions==3.7.4.2
urllib3==1.25.9
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.1.4
